remove unecessary class
